+++ src/Util/Memory.cc
@@ -6,20 +6,33 @@
 
 #include "Memory.h"
 
-#include <spdlog/fmt/fmt.h>
 #include <cerrno>
 
-bool ExcludeFromCoreDump(void* address, size_t size, std::string& message) {
+#include <spdlog/fmt/fmt.h>
+#include <spdlog/spdlog.h>
+
+/**
+ * @details Mark the memory address range provided to be excluded from core dumps, using the madvise system function, if this is supported
+ * by the platform (either MADV_DONTDUMP nor MADV_NOCORE must be defined). The address range must be page-aligned.
+ *
+ * If the functionality is unsupported or the address range is invalid, this function prints a warning message.
+ *
+ * @param address The starting address of the memory range, which must be page-aligned.
+ * @param size The size of the memory range, which must be an exact multiple of the page size.
+ *
+ * @return Whether the address range was successfully marked.
+ */
+bool ExcludeFromCoreDump(void* address, size_t size) {
 #ifdef NO_CORE_DUMP_ADVICE
     if (madvise(address, size, NO_CORE_DUMP_ADVICE)) {
         auto e(errno);
         auto out = fmt::memory_buffer();
-        fmt::format_system_error(out, e, message);
-        message = fmt::to_string(out);
+        fmt::format_system_error(out, e, "Failed to exclude data from core dump");
+        spdlog::warn(fmt::to_string(out));
         return false;
     }
     return true;
 #endif
-    message = fmt::format("{}: unsupported platform", message);
+    spdlog::warn("Could not exclude data from core dump: unsupported platform.");
     return false;
 }